# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.UserWithoutPassword:
    properties:
      avatar:
        type: string
      birthday:
        example: "2014-12-12T14:00:12+07:00"
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        description: nolint
        type: string
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyError'
      status:
        type: integer
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBody:
    properties:
      message:
        type: string
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyError:
    properties:
      error:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBody'
      status:
        type: integer
    type: object
...
  internal_models.UserWithoutID:
    properties:
      avatar:
        $ref: '#/definitions/sql.NullString'
      birthday:
        allOf:
        - $ref: '#/definitions/sql.NullTime'
        description: nolint
      email:
        type: string
      name:
        $ref: '#/definitions/sql.NullString'
      password:
        type: string
      phone:
        allOf:
        - $ref: '#/definitions/sql.NullString'
        description: nolint
    type: object
...
  internal_pkg_user_delivery.ProfileResponse:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.UserWithoutPassword'
      status:
        type: integer
    type: object
  sql.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
  sql.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: 84.23.53.28:8080
info:
  contact: {}
  description: This is a server of YULA server.
  title: YULA project API
  version: "1.0"
paths:
...
  /logout:
    post:
      description: logout in app
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: logout
      tags:
      - auth
...
  /profile/get:
    get:
      description: get profile by id
      parameters:
      - description: user id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_user_delivery.ProfileResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get profile
      tags:
      - profile
...
  /profile/update:
    patch:
      consumes:
      - application/json
      description: update some fields of profile
      parameters:
      - description: полностью опционален
        in: body
        name: user
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.UserWithoutPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_user_delivery.ProfileResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: update profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: update some fields of profile
      parameters:
      - description: полностью опционален
        in: body
        name: user
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.UserWithoutPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_user_delivery.ProfileResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: update profile
      tags:
      - profile
  /signin:
    get:
      description: signin in app
      parameters:
      - description: user email for signin
        in: query
        name: email
        required: true
        type: string
      - description: user password for signin
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: signin
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: |-
        signup in app
        Error.status can be:
        StatusErrBadRequest      = 400
        StatusErrInternalServer  = 500
      parameters:
      - description: user data for signup
        in: body
        name: preUser
        required: true
        schema:
          $ref: '#/definitions/internal_models.UserWithoutID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: signup
      tags:
      - auth
schemes:
- http
swagger: "2.0"